зробив пагінацію адекватну
