+++ src/index.js
@@ -1,11 +1,12 @@
 import './js/modal-firebase';
+import './js/displaying-movies';
 import './js/api-servise';
 import './js/refs';
 import './js/pagination';
 import './js/modal-film';
 import './js/trailer';
 import './js/modal-students';
-import './js/displaying-movies';
+
 import './js/spinner';
 import './js/preloader';
 import './js/firebase/firebaseRegAndSign';
